requirements.txt
streamlit
tensorflow
keras-nightly
Pillow
numpy